List no visible globals from check and separate by category

List no visible globals from check and separate by category

get_no_visible(path = ".", checks, ...)

Arguments

path

Path to a package tarball or a directory.

checks

Output of rcmdcheck if already computed

...

Other parameters for rcmdcheck

Value

A list with no visible globals

Examples

if (FALSE) {
# This runs a check of the example package
tempdir <- tempdir()
# Create fake package
usethis::create_package(file.path(tempdir, "checkpackage"), open = FALSE)

# Create function no visible global variables and missing documented functions
cat("
#' Function
#' @importFrom dplyr filter
#' @export
my_fun <- function() {
data %>%
filter(col == 3) %>%
mutate(new_col = 1) %>%
ggplot() +
  aes(x, y, colour = new_col) +
  geom_point()
}
", file = file.path(tempdir, "checkpackage", "R", "function.R"))

path <- file.path(tempdir, "checkpackage")
attachment::att_to_description(path = path)
get_no_visible(path)
}