Print no visible globals from check and separate by category

Print no visible globals from check and separate by category

print_globals(globals, path = ".", ...)

Arguments

globals
A list as issued from get_no_visible or empty

path
Path to a package tarball or a directory.

...
Other parameters for rcmdcheck

Examples

# NOT RUN {
tempdir <- tempdir()
# Create fake package
usethis::create_package(file.path(tempdir, "checkpackage"), open = FALSE)

# Create function no visible global variables and missing documented functions
cat("
#' Function
#' @importFrom dplyr filter
#' @export
my_fun <- function() {
data %>%
filter(col == 3) %>%
mutate(new_col = 1) %>%
ggplot() +
  aes(x, y, colour = new_col) +
  geom_point()
}
", file = file.path(tempdir, "checkpackage", "R", "function.R"))

path <- file.path(tempdir, "checkpackage")
attachment::att_to_description(path = path)
globals <- get_no_visible(path)
print_globals(globals = globals)
# }