Print no visible globals from check and separate by category

Print no visible globals from check and separate by category

print_globals(globals, path = ".", ..., message = TRUE)

Arguments

globals
A list as issued from get_no_visible or empty

path
Path to a package tarball or a directory.

...
Other parameters for rcmdcheck

message
Logical. Whether to return message with content (Default) or return as list

Examples

if (FALSE) {
tempdir <- tempdir()
# Create fake package
usethis::create_package(file.path(tempdir, "checkpackage"), open = FALSE)

# Create function no visible global variables and missing documented functions
cat("
#' Function
#' @importFrom dplyr filter
#' @export
my_fun <- function() {
data %>%
ggplot2::ggplot() +
  aes(x, y, colour = new_col) +
  geom_point()
}
", file = file.path(tempdir, "checkpackage", "R", "function.R"))

path <- file.path(tempdir, "checkpackage")
attachment::att_to_description(path = path)
globals <- get_no_visible(path)
print_globals(globals = globals)
}